+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Contract testing: Handle default response code (#431)
+
 ## v1.32.0 - (2025-07-09)
 
 - Variations: Set content-type for response